JCLOUDS-827 updating sshj
Moved sshj to the new com.hierynomus version.
This only needed a minor code change due to a method removal.

However, due to the changed groupId but the same package declarations
maven treats them as different and does not exclude the versions imported by
jsch.agentproxy so these are manually excluded and updated to the 0.0.9 versions

Also sshj is depends on a newer version of BouncyCastle. This has been updated to 1.51.
The inclusion of the extended version in the bouncycastle driver causes a duplicate
class clash. This has been excluded in the project pom.